Make new module `into_iter`.
